src/core/auth/AuthContext.tsx
import { Plugins } from '@capacitor/core';
import Axios from 'axios';
import React, { createContext, useEffect, useReducer, useState } from 'react';
import { Session } from '../models';

interface AuthState {
  session?: Session;
  loading: boolean;
  error: string;
}

const initialState: AuthState = {
  session: undefined,
  loading: false,
  error: '',
};

export type AuthAction =
  | { type: 'CLEAR_SESSION' }
  | { type: 'RESTORE_SESSION'; session: Session }
  | { type: 'LOGIN' }
  | { type: 'LOGIN_SUCCESS'; session: Session }
  | { type: 'LOGIN_FAILURE'; error: string }
  | { type: 'LOGOUT' }
  | { type: 'LOGOUT_SUCCESS' }
  | { type: 'LOGOUT_FAILURE'; error: string };

const reducer = (
  state: AuthState = initialState,
  action: AuthAction,
): AuthState => {
  switch (action.type) {
    case 'CLEAR_SESSION':
      return { ...state, session: undefined };
    case 'RESTORE_SESSION':
      return { ...state, session: action.session };
    case 'LOGIN':
      return { ...state, loading: true, error: '' };
    case 'LOGIN_SUCCESS':
      return { ...state, loading: false, session: action.session };
    case 'LOGIN_FAILURE':
      return { ...state, loading: false, error: action.error };
    case 'LOGOUT':
      return { ...state, loading: true, error: '' };
    case 'LOGOUT_SUCCESS':
      return { ...state, loading: false, session: undefined };
    case 'LOGOUT_FAILURE':
      return { ...state, loading: false, error: action.error };
    default:
      return state;
  }
};

export const AuthContext = createContext<{
  state: typeof initialState;
  dispatch: (action: AuthAction) => void;
}>({
  state: initialState,
  dispatch: () => {},
});

export const AuthProvider: React.FC = ({ children }) => {
  const [initializing, setInitializing] = useState<boolean>(true);
  const [state, dispatch] = useReducer(reducer, initialState);

  useEffect(() => {
    const { Storage } = Plugins;
    (async () => {
      const { value: token } = await Storage.get({ key: 'auth-token' });
      if (!token) return setInitializing(false);

      const headers = { Authorization: 'Bearer ' + token };
      const url = `${process.env.REACT_APP_DATA_SERVICE}/users/current`;
      const { data: user } = await Axios.get(url, { headers });

      dispatch({ type: 'RESTORE_SESSION', session: { token, user } });

      return setInitializing(false);
    })();
  }, []);

  return (
    <AuthContext.Provider value={{ state, dispatch }}>
      {initializing ? <div>Loading...</div> : children}
    </AuthContext.Provider>
  );
};
